fix typescript emiting __esModule
This is caused because whenever typescript sees an `export` or `import`,
it adds that line because it's "important for interoperability between
various loaders and transpilers". It does it even for `export type`.

I bashed my head on how to remove this for a while (almost made a custom
build step for this), but turns out if I don't export the types they are
magically added to the global scope and tsc picks them up! This even
removes some of the janky boilerplate.
See: https://github.com/microsoft/TypeScript/issues/14351#issuecomment-302257350