fix saving creations with more than 20 colors
`qs` only parses up to 20 items in an array by default; beyond that, it
parses it as an object with numbered keys instead of an array.

I've set the limit to 100 for now; I don't think we need to go beyond 55
really, but better safe than sorry.
Creations' cells are sent as an encoded string, so they're not affected